+++ gar/include/aligator/gar/dense-riccati.hpp
@@ -61,6 +61,7 @@ public:
                std::vector<VectorXs> &vs, std::vector<VectorXs> &lbdas,
                const std::optional<ConstVectorRef> &theta = std::nullopt) const;
 
+  void cycleAppend(const KnotType &knot);
   VectorRef getFeedforward(size_t i) { return datas[i].ff.matrix(); }
   RowMatrixRef getFeedback(size_t i) { return datas[i].fb.matrix(); }